at the end of targil 7

+++ src/renderer/RayTracerBasic.java
@@ -360,7 +360,7 @@ public class RayTracerBasic extends RayTracerBase {
         if (nv == 0) {
             return color;
         }
-        int nShininess = intersection.geometry.getMaterial().getnShininess();
+        int nShininess = (int) intersection.geometry.getMaterial().getnShininess();
 
         Double3 kd = intersection.geometry.getMaterial().getKd();
         Double3 ks = intersection.geometry.getMaterial().getKs();
@@ -475,8 +475,8 @@ public class RayTracerBasic extends RayTracerBase {
             return ktr;
         for (GeoPoint geopoint : intersections) {
             if(alignZero(geopoint.point.distance(gp.point)-lightDistance)<=0){
-                // if (geopoint.point.distance(gp.point) <= lightDistance &&  geopoint.geometry.getMaterial().kT.equals(new Double3(0.0))){
-                // var  kt = ktr.product(geopoint.geometry.getMaterial().kT);
+//                 if (geopoint.point.distance(gp.point) <= lightDistance &&  geopoint.geometry.getMaterial().kT.equals(new Double3(0.0))){
+//                 var  kt = ktr.product(geopoint.geometry.getMaterial().kT);
                 var kt=geopoint.geometry.getMaterial().kT;
                 ktr=kt.product(ktr);
                 if (ktr.lowerThan(MIN_CALC_COLOR_K))
